@@ -1,16 +1,16 @@
 {
-  "name": "altitudeevents/laravel-visma",
+  "name": "webparking/laravel-visma",
   "description": "Laravel API helper for Visma eAccounting",
   "license": "MIT",
   "require": {
-    "php": "^7.3",
+    "php": ">=7.4",
     "laravel/framework": "^5.5 || ^6.0 || ^7.0 || ^8.0",
     "league/oauth2-client": "^2.4",
     "ext-json": "*"
   },
   "require-dev": {
-    "friendsofphp/php-cs-fixer": "^2.16",
-    "phpstan/phpstan": "^0.12.18"
+    "friendsofphp/php-cs-fixer": "^3",
+    "phpstan/phpstan": "^1"
   },
   "autoload": {
     "psr-4": {
@@ -25,8 +25,12 @@
     }
   },
   "scripts": {
-    "php-cs-fixer": "vendor/bin/php-cs-fixer fix --config=.php_cs",
-    "test": "composer php-cs-fixer && ./vendor/bin/phpstan analyse src config --level 5"
+    "php-cs-fixer": "vendor/bin/php-cs-fixer fix --config=.php-cs-fixer.php",
+    "phpstan": "vendor/bin/phpstan analyse src config --level 5",
+    "test": [
+      "composer php-cs-fixer",
+      "composer phpstan"
+    ]
   },
   "minimum-stability": "dev",
   "prefer-stable": true
