@@ -1,26 +1,24 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Webparking\LaravelVisma\Entities;
 
 /**
- * Class VoucherRow
- *
  * @property string $AccountNumber
- * @property double $DebitAmount
- * @property double $CreditAmount
+ * @property float  $DebitAmount
+ * @property float  $CreditAmount
  * @property string $TransactionText
  * @property string $CostCenterItemId1
  * @property string $CostCenterItemId2
  * @property string $CostCenterItemId3
  * @property string $VatCodeId
  * @property string $VatCodeAndPercent
- * @property double $Quantity
- * @property double $Weight
+ * @property float  $Quantity
+ * @property float  $Weight
  * @property string $DeliveryDate
- * @property integer $HarvestYear
+ * @property int    $HarvestYear
  * @property string $ProjectId
- *
- * @package Webparking\LaravelVisma\Entities
  */
 class VoucherRow
 {
